CLN: Remove io/formats/terminal.py
Remaining is_terminal function is
used in core/config_init.py, and
as this function is not meant to
be public, we can move it there.

This will help with dependency
organization of pandas' modules.

Follow-up to gh-25886.
